users can only delete their own posts

+++ src/features/Post.js
@@ -1,32 +1,31 @@
 import { Avatar } from '@material-ui/core'
 import React from 'react'
-import db from '../firebase'
+import db, { auth } from '../firebase'
 import './Post.css'
 import ClearIcon from '@material-ui/icons/Clear';
 import { useSelector } from 'react-redux';
 import { selectChannelId } from './appSlice';
-import axios from 'axios';
 
 function Post({ user, post, timestamp, id: postId }) {
     const channelId = useSelector(selectChannelId)
 
-
+// debugger
     const handleDel = () => {
-        db.collection('channels')
-        .doc(channelId)
-        .collection('posts')
-        .doc(postId).delete().then(() => {
-            console.log("Document successfully deleted!");
-        }).catch((error) => {
-            console.error("Error removing document: ", error);
-        });
-    }
-
+        if (user.uid === auth.currentUser.uid) {
+            db.collection('channels')
+            .doc(channelId)
+            .collection('posts')
+            .doc(postId).delete().then(() => {
+                console.log("Document successfully deleted!");
+            }).catch((error) => {
+                console.error("Error removing document: ", error);
+            });
+        }
+}
 
     return (
         <div className='post'>
             <Avatar src={user.photo}
-            // onClick={handleAddChat}
             />
             <div className="post__content">
                 <h4>{user.displayName}
@@ -35,6 +34,7 @@ function Post({ user, post, timestamp, id: postId }) {
                             className='clearIcon'
                             onClick={handleDel}
                             id={post.id}
+                            user={user}
                             />
                         {new Date(timestamp?.toDate()).toUTCString()}
                     </span>
